@@ -1,22 +1,37 @@
-import React,{Component} from "react";
-import Content from "../components/content";
-import "../stylesheets/home.css";
-import { Button, Container } from "react-bootstrap";
+import {Table} from 'react-bootstrap';
 
-class Home extends Component{
-    render(){
-        var _content = 
-        <Container className="home">
-            <Button href="/manage">이수과목관리</Button>
-            <Button href="/simulation">졸업시뮬레이션</Button>
-        </Container>
-        
-        return(
-            <div>
-                <Content content={_content}></Content>
-            </div>
-        );
+function Home() {
+    function movePage(page){
+        window.location.href = page;
     }
+
+    let _contents = <Table className='Home'>
+                        <tbody>
+                            <tr>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                            </tr>
+                            <tr>
+                                <td colSpan={2} className="content" onClick={() => movePage("manage")}>이수과목관리</td>
+                                <td></td>
+                                <td colSpan={2} className="content" onClick={() => movePage("simulation")}>졸업시뮬레이션</td>
+                            </tr>
+                            <tr>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                                <td></td>
+                            </tr>
+                        </tbody>
+                    </Table>
+
+    return (
+        _contents
+    );
 }
 
-export default Home
+export default Home;
